Fix ready-to-merge reconciliation
